chore(cb2-1111): extract vin validity to be used elsewhere
