+++ tests/unit/validators/update.test.ts
@@ -1,5 +1,5 @@
 import { ERRORS, StatusCode } from '../../../src/util/enum';
-import { checkStatusCodeValidity, checkVinValidity, validateUpdateErrors } from '../../../src/validators/update';
+import { checkStatusCodeValidity, validateUpdateErrors } from '../../../src/validators/update';
 
 describe('validateUpdateErrors', () => {
   it('throws error if request body is empty', () => {
@@ -36,35 +36,5 @@ describe('checkStatusCodeValidity', () => {
     expect(checkStatusCodeValidity(StatusCode.CURRENT, StatusCode.CURRENT)).toBe(false);
     expect(checkStatusCodeValidity(StatusCode.PROVISIONAL, StatusCode.CURRENT)).toBe(false);
     expect(checkStatusCodeValidity(StatusCode.PROVISIONAL, StatusCode.PROVISIONAL)).toBe(false);
-  });
-});
-
-describe('checkVinValidity', () => {
-  it('throws error if new vin is invalid', () => {
-    expect(checkVinValidity('1234', '12')).toEqual({
-      statusCode: 400,
-      body: ERRORS.VIN_ERROR,
-    });
-    expect(checkVinValidity('1234', '123456789123456789123456789')).toEqual({
-      statusCode: 400,
-      body: ERRORS.VIN_ERROR,
-    });
-    expect(checkVinValidity('1234', '')).toEqual({
-      statusCode: 400,
-      body: ERRORS.VIN_ERROR,
-    });
-  });
-  it('should return an error if newVin contains special characters', () => {
-    const result = checkVinValidity('12345', '!newvin');
-    expect(result).toEqual({
-      statusCode: 400,
-      body: ERRORS.VIN_ERROR,
-    });
-  });
-  it('returns false if no errors', () => {
-    expect(checkVinValidity('1234', '1234')).toBe(false);
-    expect(checkVinValidity('1234', '12345')).toBe(false);
-    expect(checkVinValidity('1234', undefined)).toBe(false);
-    expect(checkVinValidity('1234', null)).toBe(false);
   });
 });
